Delete old HTTP methods
